i#19: use cmake as pre-make configure step
new features from using cmake:
* configuration of build to use local toolchain: moved build-time checks
  for -dT, -fvisibility, etc. to configure-time
* core, api/docs, libutil, tools all tied together in one build system
  - user-friendly top-level control over what to build
  - inter-module dependencies
    no more manual "make DEBUG=0 api_headers" and then manual
    docs creation after tweaking header files
  - libutil and tools have exact defines used in core, from configure.h
* doxygen warnings on missing references are now fatal errors
* proper dependencies in all modules
* make -j support
* rely on fewer toolchain utilities:
  - cygwin no longer required!
    for core and tools just need any perl; for docs need additional
    tools but do not have to be cygwin
  - replaced sed, awk, grep, cat, cut, head, tail, wc with cmake scripts
  - i#63: replace tar and zip with CPack
  - i#68/i#65: replace perl with cmake scripts and CTest
* docs building gracefully optional

docs changes:
* removed "$relpath$" from api/docs/header.html as it is only for
  CREATE_SUBDIRS=yes, and new doxygen 1.5.7+ doesn't expand it otherwise
* removed support for VMSAFE docs
* removed latex target and pdf rules: i#59 covers re-adding if desired

core changes:
* removed support for NOLIBC=0 Windows builds
* no longer building lib/fortran.c -- we weren't exporting it anyway,
  and few people use the start/stop interface
* moved generated defines.h DYNAMORIO_DEFINES to configure.h
  and buildmark.c externs to globals.h
  => file can't have '#ifdef FOO' around whole thing: has to include
     configure.h first
* moved dr_api.h and dr_app.h copy from core/Makefile to genapi.pl;
  eliminated dr_app.h complex processing since no longer needed
* i#18: check for uint, etc. typedef conflicts like on rhel3 
  and set DR_DO_NOT_DEFINE_*
* fixed build errors on rhel3:
  - #defines inside macros in linux/os.c and linux/signal.c
  - still hit issues filed as i#TOFILE: 
    RHEL3 linker script has no __executable_start: I tried to add it
    but not working 

libutil changes:
* policy_staticd is no more: simply policy_static w/ debug config
* including configure.h
* removed tests rules: i#60 covers re-adding and getting to work again

all changes:
* moved FILE_* defines into resources.rc since hard to pass spaces to nmake
  through cmake: but our key files already had names as #defines

not yet finished:
* update wiki/HowToBuild.wiki
i#65: suite/; then remove make/compiler.mk
i#63: packaging; remove batch/* and makezip.sh afterward
i#67: api/samples/Makefile
i#64: "make review"
i#66: "make runregression" + internal build#; then remove make/custom_build.mk

future nice-to-have work:
i#68: replace perl with cmake scripts and CTest
i#69: re-add support for core unit tests and standalone build rules
i#70: symbol store support in core/Makefile and tools/Makefile
i#71: support mingw/cygwin makefile cmake target with cl compiler, for make -j


SVN-Revision: 64

+++ tools/runstats.c
@@ -97,7 +97,7 @@ get_mem_stats(pid_t pid)
     if (status == NULL)
         return; /* child already exited */
     while(!feof(status)){
-        unsigned long val = 0;
+        unsigned val = 0;
         char prefix[STATUS_LINE_LENGTH];
         int len;
         if (NULL==fgets(line, sizeof(line), status))
